+++ packages/press-ui/src/packages/mixins/click-outside.js
@@ -29,3 +29,5 @@ export const ClickOutsideMixin = config => ({
     document.removeEventListener(config.event, this.clickOutsideHandler);
   },
 });
+
+export const clickOutsideMixin = ClickOutsideMixin;